New AppConfigurationClient trait + IBMCloud implementation
