+++ src/tests/test_using_example_data.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::client::AppConfigurationClient;
+use crate::client::{AppConfigurationClientIBMCloud, AppConfigurationClient};
 use crate::tests::TrivialEntity;
 use rstest::*;
 
@@ -20,7 +20,7 @@ use super::client_enterprise;
 use crate::{Feature, Property, Value};
 
 #[rstest]
-fn test_get_a_specific_feature(client_enterprise: AppConfigurationClient) {
+fn test_get_a_specific_feature(client_enterprise: AppConfigurationClientIBMCloud) {
     let specific_feature = client_enterprise.get_feature_proxy("f1").unwrap();
 
     let name = specific_feature.get_name().unwrap();
@@ -33,7 +33,7 @@ fn test_get_a_specific_feature(client_enterprise: AppConfigurationClient) {
 }
 
 #[rstest]
-fn test_get_a_specific_property(client_enterprise: AppConfigurationClient) {
+fn test_get_a_specific_property(client_enterprise: AppConfigurationClientIBMCloud) {
     let property = client_enterprise.get_property_proxy("p1").unwrap();
 
     let name = property.get_name().unwrap();
